Getting up to date

+++ src/vm/fcall.cpp
@@ -1,7 +1,6 @@
-//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-//
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
 // FCALL.CPP
 //
 
@@ -28,7 +27,7 @@ NOINLINE LPVOID __FCThrow(LPVOID __me, RuntimeExceptionKind reKind, UINT resID, 
 
     // side effect the compiler can't remove
     if (FC_NO_TAILCALL != 1)
-        return (LPVOID)(FC_NO_TAILCALL + 1);
+        return (LPVOID)(SIZE_T)(FC_NO_TAILCALL + 1);
 
     FC_CAN_TRIGGER_GC();
     INCONTRACT(FCallCheck __fCallCheck(__FILE__, __LINE__));
@@ -72,7 +71,7 @@ NOINLINE LPVOID __FCThrowArgument(LPVOID __me, RuntimeExceptionKind reKind, LPCW
 
     // side effect the compiler can't remove
     if (FC_NO_TAILCALL != 1)
-        return (LPVOID)(FC_NO_TAILCALL + 1);
+        return (LPVOID)(SIZE_T)(FC_NO_TAILCALL + 1);
 
     FC_CAN_TRIGGER_GC();
     INCONTRACT(FCallCheck __fCallCheck(__FILE__, __LINE__));
